Make the result of Pool::run Send.
Without this, it is challening to use the connection pool with Hyper. Hyper expects all of its futures to be Send, so to use the pool, a Send future would need to wait on a non-Send future. That may be possible, but I have no idea how to do it.