Add connection pool for Redis
This creates a RedisPool and RedisConnectionManager type compatible with bb8.
The RedisPool struct exists to paper over the small API incompatibility between
the redis and tokio-postgres crates. While the tokio-postgres API always
returns the connection, the redis API only returns the connection on success.
The RedisPool hides that by creating a RedisConnectionManager that uses a
Option<Connection> and returns None on failure. There's also a smaller
incompatibility in that tokio-postgres uses (T, Connection) and redis uses
(Connection, T).