Implement non-retriable errors
`ManageConnection::should_retry` can now be implemented to prevent bb8
from attempting to retry connection attempts. This serves two purposes:

1. The caller of `pool.get` will now receive the underlying exception if
a "catastrophic" error is hit. Previously, the caller would get a
`TimedOut` with no cause.

2. The error can be surfaced quickly, rather than going through several
connection attempts.

This is particularly useful for interactive use of bb8, such as allowing
a user to directly input a connection to a program such a CLI. If the
connection is incorrect, it is desirable to fail fast with a useful
error message.