Add reap_all_idle_connections option, fixes #123
Adds a new boolean option to Builder named `reap_all_idle_connections`. If true (the default value), the current behavior of the pool is maintained. However, if false, the reaper will not close any idle connections if it results in there being fewer than `min_idle` connections, even if the connections have been idle for longer than `idle_timeout`.

This option would best be disabled if connections are expensive to establish and/or are generally stable over time.