Change notify_waiters to notify_one on broken conns
When a connection is dropped but broken, we notify all waiting
get requests that they should try again. This is done in order
to ensure we have enough pending connections to serve all in-
flight gets. Because only a single connection dropped however,
we only need to notify one waiter to refresh, not all of them.